refactor: nest component files and update imports
