+++ ui-library/components/index.ts
@@ -1,0 +1,11 @@
+export { default as BaseButton } from './BaseButton/BaseButton.vue';
+export { default as BaseCheckbox } from './BaseCheckbox/BaseCheckbox.vue';
+export { default as BaseCollapse } from './BaseCollapse/BaseCollapse.vue';
+export { default as BaseDropdown } from './BaseDropdown/BaseDropdown.vue';
+export { default as BaseInput } from './BaseInput/BaseInput.vue';
+export { default as BaseModal } from './BaseModal/BaseModal.vue';
+export { default as BaseSelect } from './BaseSelect/BaseSelect.vue';
+export { default as BaseSwitch } from './BaseSwitch/BaseSwitch.vue';
+export { default as BaseTab } from './BaseTab/BaseTab.vue';
+export { default as BaseTextarea } from './BaseTextarea/BaseTextarea.vue';
+export { default as BaseTooltip } from './BaseTooltip/BaseTooltip.vue';
