+++ ui-library/index.ts
@@ -1,7 +1,1 @@
-export { default as BaseButton } from './components/BaseButton.vue';
-export { default as BaseSwitch } from './components/BaseSwitch.vue';
-export { default as BaseModal } from './components/BaseModal.vue';
-export { default as BaseTextarea } from './components/BaseTextarea.vue';
-export { default as BaseSelect } from './components/BaseSelect.vue';
-export { default as BaseTooltip } from './components/BaseTooltip.vue';
-export { default as BaseTab } from './components/BaseTab.vue';
+export * from './components';
