+++ ui-library/types/index.d.ts
@@ -1,1 +1,11 @@
-export { default as BaseButton } from '../components/BaseButton.vue';
+export { default as BaseButton } from '../components/BaseButton/BaseButton.vue';
+export { default as BaseCheckbox } from '../components/BaseCheckbox/BaseCheckbox.vue';
+export { default as BaseCollapse } from '../components/BaseCollapse/BaseCollapse.vue';
+export { default as BaseDropdown } from '../components/BaseDropdown/BaseDropdown.vue';
+export { default as BaseInput } from '../components/BaseInput/BaseInput.vue';
+export { default as BaseModal } from '../components/BaseModal/BaseModal.vue';
+export { default as BaseSelect } from '../components/BaseSelect/BaseSelect.vue';
+export { default as BaseSwitch } from '../components/BaseSwitch/BaseSwitch.vue';
+export { default as BaseTab } from '../components/BaseTab/BaseTab.vue';
+export { default as BaseTextarea } from '../components/BaseTextarea/BaseTextarea.vue';
+export { default as BaseTooltip } from '../components/BaseTooltip/BaseTooltip.vue';
